completed

+++ src/controllers/CustomerChannelController.php
@@ -1,13 +1,12 @@
 <?php
 
 namespace Abs\CustomerChannelPkg;
-use Abs\CustomerChannelGroupPkg\CustomerChannelGroup;
-use App\Address;
-use App\Country;
+use Abs\CustomerChannelPkg\CustomerChannelGroup;
 use App\Http\Controllers\Controller;
 use Auth;
 use Carbon\Carbon;
 use DB;
+use Entrust;
 use Illuminate\Http\Request;
 use Validator;
 use Yajra\Datatables\Datatables;
@@ -21,26 +20,13 @@ class CustomerChannelGroupController extends Controller {
 		$customer_channel_groups = CustomerChannelGroup::withTrashed()
 			->select(
 				'customer_channel_groups.id',
-				'customer_channel_groups.code',
 				'customer_channel_groups.name',
-				DB::raw('IF(customer_channel_groups.mobile_no IS NULL,"--",customer_channel_groups.mobile_no) as mobile_no'),
-				DB::raw('IF(customer_channel_groups.email IS NULL,"--",customer_channel_groups.email) as email'),
 				DB::raw('IF(customer_channel_groups.deleted_at IS NULL,"Active","Inactive") as status')
 			)
 			->where('customer_channel_groups.company_id', Auth::user()->company_id)
 			->where(function ($query) use ($request) {
-				if (!empty($request->customer_channel_group_code)) {
-					$query->where('customer_channel_groups.code', 'LIKE', '%' . $request->customer_channel_group_code . '%');
-				}
-			})
-			->where(function ($query) use ($request) {
-				if (!empty($request->customer_channel_group_name)) {
-					$query->where('customer_channel_groups.name', 'LIKE', '%' . $request->customer_channel_group_name . '%');
-				}
-			})
-			->where(function ($query) use ($request) {
-				if (!empty($request->mobile_no)) {
-					$query->where('customer_channel_groups.mobile_no', 'LIKE', '%' . $request->mobile_no . '%');
+				if (!empty($request->name)) {
+					$query->where('customer_channel_groups.name', 'LIKE', '%' . $request->name . '%');
 				}
 			})
 			->where(function ($query) use ($request) {
@@ -48,25 +34,34 @@ class CustomerChannelGroupController extends Controller {
 					$query->where('customer_channel_groups.email', 'LIKE', '%' . $request->email . '%');
 				}
 			})
+			->whereNull('customer_channel_groups.parent_id')
 			->orderby('customer_channel_groups.id', 'desc');
 
 		return Datatables::of($customer_channel_groups)
-			->addColumn('code', function ($customer_channel_group) {
-				$status = $customer_channel_group->status == 'Active' ? 'green' : 'red';
-				return '<span class="status-indicator ' . $status . '"></span>' . $customer_channel_group->code;
-			})
+		// ->addColumn('code', function ($customer_channel_group) {
+		// 	$status = $customer_channel_group->status == 'Active' ? 'green' : 'red';
+		// 	return '<span class="status-indicator ' . $status . '"></span>' . $customer_channel_group->code;
+		// })
 			->addColumn('action', function ($customer_channel_group) {
-				$edit_img = asset('public/theme/img/table/cndn/edit.svg');
-				$delete_img = asset('public/theme/img/table/cndn/delete.svg');
-				return '
-					<a href="#!/customer_channel_group-pkg/customer_channel_group/edit/' . $customer_channel_group->id . '">
-						<img src="' . $edit_img . '" alt="View" class="img-responsive">
-					</a>
-					<a href="javascript:;" data-toggle="modal" data-target="#delete_customer_channel_group"
+				$edit = asset('public/img/content/table/edit-yellow.svg');
+				$edit_active = asset('public/img/content/table/edit-yellow-active.svg');
+				$delete = asset('/public/img/content/table/delete-default.svg');
+				$delete_active = asset('/public/img/content/table/delete-active.svg');
+
+				$action = '';
+				if (Entrust::can('edit-customer-channel-group')) {
+					$action .= '<a href="#!/customer-channel-pkg/customer-channel-group/edit/' . $customer_channel_group->id . '">
+						<img src="' . $edit . '" alt="Edit" class="img-responsive" onmouseover=this.src="' . $edit_active . '" onmouseout=this.src="' . $edit . '" >
+					</a>';
+				}
+				if (Entrust::can('delete-customer-channel-group')) {
+					$action .= '<a href="javascript:;" data-toggle="modal" data-target="#delete_customer_channel_group"
 					onclick="angular.element(this).scope().deleteCustomerChannelGroup(' . $customer_channel_group->id . ')" dusk = "delete-btn" title="Delete">
-					<img src="' . $delete_img . '" alt="delete" class="img-responsive">
+					<img src="' . $delete . '" alt="Delete" class="img-responsive" onmouseover=this.src="' . $delete_active . '" onmouseout=this.src="' . $delete . '" >
 					</a>
 					';
+				}
+				return $action;
 			})
 			->make(true);
 	}
@@ -74,19 +69,15 @@ class CustomerChannelGroupController extends Controller {
 	public function getCustomerChannelGroupFormData($id = NULL) {
 		if (!$id) {
 			$customer_channel_group = new CustomerChannelGroup;
-			$address = new Address;
+			$customer_channel_sub_group = [];
 			$action = 'Add';
 		} else {
 			$customer_channel_group = CustomerChannelGroup::withTrashed()->find($id);
-			$address = Address::where('address_of_id', 24)->where('entity_id', $id)->first();
-			if (!$address) {
-				$address = new Address;
-			}
+			$customer_channel_sub_group = CustomerChannelGroup::withTrashed()->where('parent_id', $id)->get();
 			$action = 'Edit';
 		}
-		$this->data['country_list'] = $country_list = Collect(Country::select('id', 'name')->get())->prepend(['id' => '', 'name' => 'Select Country']);
 		$this->data['customer_channel_group'] = $customer_channel_group;
-		$this->data['address'] = $address;
+		$this->data['customer_channel_sub_group'] = $customer_channel_sub_group;
 		$this->data['action'] = $action;
 
 		return response()->json($this->data);
@@ -96,57 +87,61 @@ class CustomerChannelGroupController extends Controller {
 		// dd($request->all());
 		try {
 			$error_messages = [
-				'code.required' => 'CustomerChannelGroup Code is Required',
-				'code.max' => 'Maximum 255 Characters',
-				'code.min' => 'Minimum 3 Characters',
-				'code.unique' => 'CustomerChannelGroup Code is already taken',
-				'name.required' => 'CustomerChannelGroup Name is Required',
+				'name.required' => 'Customer Channel Group Name is Required',
+				'name.unique' => 'Customer Channel Group Name is already taken',
 				'name.max' => 'Maximum 255 Characters',
 				'name.min' => 'Minimum 3 Characters',
-				'gst_number.required' => 'GST Number is Required',
-				'gst_number.max' => 'Maximum 191 Numbers',
-				'mobile_no.max' => 'Maximum 25 Numbers',
-				// 'email.required' => 'Email is Required',
-				'address_line1.required' => 'Address Line 1 is Required',
-				'address_line1.max' => 'Maximum 255 Characters',
-				'address_line1.min' => 'Minimum 3 Characters',
-				'address_line2.max' => 'Maximum 255 Characters',
-				// 'pincode.required' => 'Pincode is Required',
-				// 'pincode.max' => 'Maximum 6 Characters',
-				// 'pincode.min' => 'Minimum 6 Characters',
 			];
 			$validator = Validator::make($request->all(), [
-				'code' => [
+				'name' => [
 					'required:true',
 					'max:255',
 					'min:3',
-					'unique:customer_channel_groups,code,' . $request->id . ',id,company_id,' . Auth::user()->company_id,
+					'unique:customer_channel_groups,name,' . $request->id . ',id,company_id,' . Auth::user()->company_id . ',parent_id,NULL',
 				],
-				'name' => 'required|max:255|min:3',
-				'gst_number' => 'required|max:191',
-				'mobile_no' => 'nullable|max:25',
-				// 'email' => 'nullable',
-				'address' => 'required',
-				'address_line1' => 'required|max:255|min:3',
-				'address_line2' => 'max:255',
-				// 'pincode' => 'required|max:6|min:6',
 			], $error_messages);
 			if ($validator->fails()) {
 				return response()->json(['success' => false, 'errors' => $validator->errors()->all()]);
 			}
 
+			$error_messages1 = [
+				'sub_group_name.required' => 'Customer Channel Group Name is Required',
+				'sub_group_name.unique' => 'Customer Channel Sub Group Name is already taken',
+				'sub_group_name.max' => 'Maximum 255 Characters',
+				'sub_group_name.min' => 'Minimum 3 Characters',
+			];
+			if (!empty($request->customer_channel_groups)) {
+				foreach ($request->customer_channel_groups as $customer_channel_group) {
+					$validator = Validator::make($customer_channel_group, [
+						'sub_group_name' => [
+							'required:true',
+							'max:191',
+							'min:3',
+							'unique:customer_channel_groups,name,' . $customer_channel_group['id'] . ',id,company_id,' . Auth::user()->company_id . ',parent_id,' . $request->id,
+						],
+					], $error_messages1);
+					if ($validator->fails()) {
+						return response()->json(['success' => false, 'errors' => $validator->errors()->all()]);
+					}
+				}
+			}
+
 			DB::beginTransaction();
+			// dd($request->customer_channel_sub_group_id);
+			if (!empty($request->customer_channel_sub_group_id)) {
+				$channel_group_removal_id = json_decode($request->customer_channel_sub_group_id, true);
+				CustomerChannelGroup::whereIn('id', $channel_group_removal_id)->forceDelete();
+			}
+
 			if (!$request->id) {
 				$customer_channel_group = new CustomerChannelGroup;
 				$customer_channel_group->created_by_id = Auth::user()->id;
 				$customer_channel_group->created_at = Carbon::now();
 				$customer_channel_group->updated_at = NULL;
-				$address = new Address;
 			} else {
 				$customer_channel_group = CustomerChannelGroup::withTrashed()->find($request->id);
 				$customer_channel_group->updated_by_id = Auth::user()->id;
 				$customer_channel_group->updated_at = Carbon::now();
-				$address = Address::where('address_of_id', 24)->where('entity_id', $request->id)->first();
 			}
 			$customer_channel_group->fill($request->all());
 			$customer_channel_group->company_id = Auth::user()->company_id;
@@ -157,26 +152,39 @@ class CustomerChannelGroupController extends Controller {
 				$customer_channel_group->deleted_by_id = NULL;
 				$customer_channel_group->deleted_at = NULL;
 			}
-			$customer_channel_group->gst_number = $request->gst_number;
-			$customer_channel_group->axapta_location_id = $request->axapta_location_id;
 			$customer_channel_group->save();
 
-			if (!$address) {
-				$address = new Address;
+			if (!empty($request->customer_channel_groups)) {
+				foreach ($request->customer_channel_groups as $customer_channel_groups) {
+					if (!$customer_channel_groups['id']) {
+						$customer_channel_sub_group = new CustomerChannelGroup;
+						$customer_channel_sub_group->parent_id = $customer_channel_group->id;
+						$customer_channel_sub_group->created_by_id = Auth::user()->id;
+						$customer_channel_sub_group->created_at = Carbon::now();
+						$customer_channel_sub_group->updated_at = NULL;
+					} else {
+						$customer_channel_sub_group = CustomerChannelGroup::withTrashed()->find($customer_channel_groups['id']);
+						$customer_channel_sub_group->updated_by_id = Auth::user()->id;
+						$customer_channel_sub_group->updated_at = Carbon::now();
+					}
+					$customer_channel_sub_group->name = $customer_channel_groups['sub_group_name'];
+					$customer_channel_sub_group->company_id = Auth::user()->company_id;
+					if ($customer_channel_groups['status'] == 'Inactive') {
+						$customer_channel_sub_group->deleted_at = Carbon::now();
+						$customer_channel_sub_group->deleted_by_id = Auth::user()->id;
+					} else {
+						$customer_channel_sub_group->deleted_by_id = NULL;
+						$customer_channel_sub_group->deleted_at = NULL;
+					}
+					$customer_channel_sub_group->save();
+				}
 			}
-			$address->fill($request->all());
-			$address->company_id = Auth::user()->company_id;
-			$address->address_of_id = 24;
-			$address->entity_id = $customer_channel_group->id;
-			$address->address_type_id = 40;
-			$address->name = 'Primary Address';
-			$address->save();
 
 			DB::commit();
 			if (!($request->id)) {
-				return response()->json(['success' => true, 'message' => ['CustomerChannelGroup Details Added Successfully']]);
+				return response()->json(['success' => true, 'message' => ['Customer Channel Group Details Added Successfully']]);
 			} else {
-				return response()->json(['success' => true, 'message' => ['CustomerChannelGroup Details Updated Successfully']]);
+				return response()->json(['success' => true, 'message' => ['Customer Channel Group Details Updated Successfully']]);
 			}
 		} catch (Exceprion $e) {
 			DB::rollBack();
@@ -184,9 +192,9 @@ class CustomerChannelGroupController extends Controller {
 		}
 	}
 	public function deleteCustomerChannelGroup($id) {
-		$delete_status = CustomerChannelGroup::withTrashed()->where('id', $id)->forceDelete();
-		if ($delete_status) {
-			$address_delete = Address::where('address_of_id', 24)->where('entity_id', $id)->forceDelete();
+		$sub_group_delete_status = CustomerChannelGroup::withTrashed()->where('parent_id', $id)->forceDelete();
+		if ($sub_group_delete_status) {
+			$delete_status = CustomerChannelGroup::withTrashed()->where('id', $id)->forceDelete();
 			return response()->json(['success' => true]);
 		}
 	}
